added plantUML diagram. Added GlobalExceptionHandler. Added Shop entity. Computer controller and service modifications. Updated data.sql with a data insertion to 'shop' table. Added tests. Project tested with 86% code coverage.

+++ src/main/java/org/factoriaf5/computers_shop/services/ComputerService.java
@@ -5,7 +5,6 @@ import java.util.List;
 import org.factoriaf5.computers_shop.models.Computer;
 import org.factoriaf5.computers_shop.repositories.ComputerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -18,20 +17,24 @@ public class ComputerService {
     }
 
     public Computer create(Computer computer){
-        try{
-            return repository.save(computer);
-        }catch(DataIntegrityViolationException e){
-            return null;
-        }
+        return repository.save(computer);
     }
 
     public List<Computer> findByBrand(String brand){
+        if(brand == null){
+            throw new IllegalArgumentException();
+        }
         return repository.findByBrand(brand);
     }
 
     public Integer deleteByBrand(String brand){
+        if(brand == null){
+            throw new IllegalArgumentException();
+        }
         List<Computer> computers = repository.findByBrand(brand);
-        repository.deleteAll(computers);
+        if(!computers.isEmpty()){
+            repository.deleteAll(computers);
+        }
         return computers.size();
     }
